requirements.txt
fastapi
uvicorn[standard]
langgraph>=1.0.0
langchain-core
google-genai
httpx
python-dotenv
pydantic
